split up crm tests
